cic-eth: Revert number as hex in tx-cache

See merge request grassrootseconomics/cic-internal-integration!31

+++ apps/cic-eth/cic_eth/db/migrations/postgresql/versions/e3b5330ee71c_.py
@@ -19,7 +19,6 @@ def upgrade():
     op.create_table(
             'tx_cache',
             sa.Column('id', sa.Integer, primary_key=True),
-#            sa.Column('tx_id', sa.Integer, sa.ForeignKey('tx.id'), nullable=True),
             sa.Column('otx_id', sa.Integer, sa.ForeignKey('otx.id'), nullable=True),
             sa.Column('date_created', sa.DateTime, nullable=False),
             sa.Column('date_updated', sa.DateTime, nullable=False),
@@ -27,8 +26,8 @@ def upgrade():
             sa.Column('destination_token_address', sa.String(42), nullable=False),
             sa.Column('sender', sa.String(42), nullable=False),
             sa.Column('recipient', sa.String(42), nullable=False),
-            sa.Column('from_value', sa.BIGINT(), nullable=False),
-            sa.Column('to_value', sa.BIGINT(), nullable=True),
+            sa.Column('from_value', sa.NUMERIC(), nullable=False),
+            sa.Column('to_value', sa.NUMERIC(), nullable=True),
             sa.Column('block_number', sa.BIGINT(), nullable=True),
             sa.Column('tx_index', sa.Integer, nullable=True),
             )
